Words in text map creation
First work in creating a map of words in the texts. The map keys are the
words, and their value is an integer which tells the number of
occurrences in the text for the word.

Need to organize Texts.h and start working on a base for the inverted
index.