Complete texts scan and token data class.
Completed the text scanning and word retrieving functionality. Still
some bugs regarding words with numbers, but will leave it like that for
now.

Added a class for token (word found) data. This will be used to know if
a word is found in a text, how many times, and at what lines. Included
some testing for the class too.